fix api routes
