+++ lib/routes/index.js
@@ -12,7 +12,13 @@ router.get('/', function(req, res, next) {
   });
 });
 
-router.get('*', function(req, res) {
+router.get('/v1', function(req, res, next) {
+  res.send({
+    apiVersion : 1.0
+  });
+});
+
+router.get('/*', function(req, res) {
   res.status(404)        // HTTP status 404: NotFound
       .send('Not found');
 });
