+++ lib/routes/status.js
@@ -53,15 +53,14 @@ function update (req, res) {
 }
 
 // routes
-
-router.get('/status', (req, res) => {
+router.get('/', (req, res) => {
   req.params.hash = req.query.d
   show(req, res)
 })
 
-router.get('/status/:hash', show)
+router.get('/:hash', show)
 
-router.post('/status', (req, res) => {
+router.post('/', (req, res) => {
   req.params.hash = req.body.hash || req.body.d
   update(req, res)
 })
