--- lib/server.js
@@ -1,21 +1,0 @@
-'use strict'
-
-var config = require('config')
-var db = require('./db')
-
-const app = require('./web')
-
-var server = app.listen(config.get('app.port'), function () {
-  var host = server.address().address
-  var port = server.address().port
-
-  console.log('Example app listening at http://%s:%s', host, port)
-})
-
-function stop () {
-  server.close()
-  db.close()
-}
-
-module.exports = server
-module.exports.stop = stop
